# Speed up the lexer for Ruby 3.4+