--- a/.github/workflows/liquid.yml
+++ b/.github/workflows/liquid.yml
@@ -11,7 +11,7 @@ jobs:
     strategy:
       matrix:
         entry:
-          - { ruby: 2.7, allowed-failure: false } # minimum supported
+          - { ruby: 3.0, allowed-failure: false } # minimum supported
           - { ruby: 3.2, allowed-failure: false }
           - { ruby: 3.3, allowed-failure: false } # latest
           - { ruby: ruby-head, allowed-failure: true }