--- a/performance/benchmark.rb
+++ b/performance/benchmark.rb
@@ -3,6 +3,7 @@
 require 'benchmark/ips'
 require_relative 'theme_runner'
 
+RubyVM::YJIT.enable if defined?(RubyVM::YJIT)
 Liquid::Template.error_mode = ARGV.first.to_sym if ARGV.first
 profiler = ThemeRunner.new