… .ruby-version
… Gemfile.lock