This PR speeds up lexing by making using of some new `strscan` features when available, namely `StringScanner#peek_byte`.

Comparison of `be rake benchmark:strict` on Ruby 3.4:
- **old parse**: `72.452 (± 4.1%) i/s   (13.80 ms/i) -      2.170k in  30.029157s`
- **new parse**: `92.409 (± 1.1%) i/s   (10.82 ms/i) -      2.772k in  30.002144s`

To make this work smoothly with CI and such, I needed to bump min version to Ruby 3.0 and remove Gemfile.lock, which need to be updated anyway.